Make enum abstracts to reduce string allocation (Big commit)
Character animation names, hscript function names, and the strumnote animation names are now enum abstracts so that string allocation isn't so excessive in-game (Especially with hscript), so this is a memory-wise optimization you can kind of notice.

Also, I'll deal with the crash a few moments after the song ends later or so.