Implement binary format (Yet another big commit)
Yes I actually did it. I also revamped the song path system to be fancier.
I even made byte streaming, and that is also an option. It was also a pain in the ass tft because I faced errors that came out of nowhere... Then, I finally fixed that error, made the binary chart format parser fully stable, and got note spawning working in no time at all.
It was really fun to do.